BE/requirements.txt
fastapi==0.85.0
pydantic==1.10.2
uvicorn==0.19.0
psycopg2-binary==2.8.6
asyncpg==0.22.0
SQLAlchemy==1.4.35
passlib==1.7.4
PyJWT==2.6.0
python-jose==3.3.0